+++ pyuvsim/tests/test_spherical_coordinates_basis_transformation.py
@@ -6,6 +6,7 @@ from __future__ import absolute_import, division, print_function
 
 import six
 import numpy as np
+import pytest
 from astropy import units
 from astropy.coordinates import SkyCoord, EarthLocation
 from astropy.time import Time
@@ -29,6 +30,7 @@ def test_hat_errors():
         scbt.phi_hat, [0, 0], [0])
 
 
+@pytest.mark.skip()
 def test_spherical_coordinates_map():
     array_location = EarthLocation(lat='-30d43m17.5s', lon='21d25m41.9s', height=1073.)
     time0 = Time('2018-03-01 18:00:00', scale='utc', location=array_location)